# Replace isarray and isobject with built-in js methods