--- a/lib/line-column.js
+++ b/lib/line-column.js
@@ -5,10 +5,6 @@
  */
 "use strict";
 
-var isArray  = require("isarray");
-var isObject = require("isobject");
-var slice = Array.prototype.slice;
-
 module.exports = LineColumnFinder;
 
 /**
@@ -72,10 +68,10 @@ LineColumnFinder.prototype.fromIndex = function (index) {
  */
 LineColumnFinder.prototype.toIndex = function (line, column) {
   if (typeof column === "undefined") {
-    if (isArray(line) && line.length >= 2) {
+    if (Array.isArray(line) && line.length >= 2) {
       return this.toIndex(line[0], line[1]);
     }
-    if (isObject(line) && "line" in line && ("col" in line || "column" in line)) {
+    if (line != null && typeof line === 'object' && "line" in line && ("col" in line || "column" in line)) {
       return this.toIndex(line.line, ("col" in line ? line.col : line.column));
     }
     return -1;