--- a/package.json
+++ b/package.json
@@ -40,8 +40,5 @@
     "mocha": "^2.4.5",
     "power-assert": "^1.3.1"
   },
-  "dependencies": {
-    "isarray": "^1.0.0",
-    "isobject": "^2.0.0"
-  }
+  "dependencies": {}
 }