`isarray` and `isobject` are no longer required in the current state of JavaScript as there are native alternatives. Removing those dependencies can help to reduce unnecessary network traffic when this package is being installed.

Implemented testing and benchmark to make sure that the change doesn't break anything and the performance difference is within margin of error.

**Before**
```
long text  + line-column (not cached) x 117,262 ops/sec ±0.34% (87 runs sampled)
long text  + line-column (cached) x 12,636,686 ops/sec ±1.41% (90 runs sampled)
long text  + find-line-column x 59,024 ops/sec ±0.16% (99 runs sampled)
short text + line-column (not cached) x 3,438,077 ops/sec ±0.17% (97 runs sampled)
short text + line-column (cached) x 24,706,656 ops/sec ±1.50% (93 runs sampled)
short text + find-line-column x 1,638,901 ops/sec ±0.09% (98 runs sampled)
```
**After**
```
long text  + line-column (not cached) x 113,826 ops/sec ±2.28% (90 runs sampled)
long text  + line-column (cached) x 12,805,129 ops/sec ±0.45% (96 runs sampled)
long text  + find-line-column x 58,999 ops/sec ±0.18% (97 runs sampled)
short text + line-column (not cached) x 3,432,124 ops/sec ±0.18% (97 runs sampled)
short text + line-column (cached) x 24,947,691 ops/sec ±0.99% (89 runs sampled)
short text + find-line-column x 1,635,250 ops/sec ±0.13% (98 runs sampled)
```